Fix Makefile syntax and address remaining CI/PR issues
- Fixed Makefile syntax error "recipe commences before first target".
- Fixed Ruff linting errors (import sorting, unused imports, single-line colons).
- Fixed Mypy type checking errors (missing type annotations, optional types).
- Implemented robust project root discovery in `HiveCell`.
- Refactored `TelegramEffector` for robust `chat_id` extraction.
- Switched to structured logging with `exc_info=True`.
- Replaced remaining "adapters" with "synapses" in code comments.
- Added `__init__.py` files to proto directories to help Mypy.
- Verified system integrity with all 54 core tests and bot tests passing.
