+++ Makefile
@@ -6,7 +6,7 @@ REGISTRY ?= ghcr.io/zaebee
 PLATFORM ?= linux/amd64
 CORE_PATH ?= core:core/src
 GATEWAY_PATH ?= api-gateway/src
-TG_PATH ?= synapses/telegram-bot/src:synapses/telegram-bot/src/proto
+TG_PATH ?= synapses/telegram-bot:synapses/telegram-bot/proto
 MCP_PATH ?= synapses/mcp-server:synapses/mcp-server/proto
 KEEPER_PATH ?= agents/bee-keeper/src
 
@@ -19,8 +19,8 @@ lint:
 	# Python Type Check (Mypy)
 	MYPYPATH=$(CORE_PATH) uv run mypy core/src
 	MYPYPATH=$(GATEWAY_PATH):packages/aura-core/src uv run mypy api-gateway/src
-	MYPYPATH=$(TG_PATH):packages/aura-core/src uv run mypy synapses/telegram-bot/src
-  MYPYPATH=$(MCP_PATH):packages/aura-core/src uv run mypy synapses/mcp-server
+	MYPYPATH=$(TG_PATH):packages/aura-core/src uv run mypy synapses/telegram-bot
+	MYPYPATH=$(MCP_PATH):packages/aura-core/src uv run mypy synapses/mcp-server
 	MYPYPATH=$(KEEPER_PATH):packages/aura-core/src uv run mypy agents/bee-keeper/main.py agents/bee-keeper/src
 	MYPYPATH=packages/aura-core/src uv run mypy packages/aura-core/src
 	# Security Audit (Bandit)
@@ -37,15 +37,15 @@ test:
 	# Run core tests
 	PYTHONPATH=$(CORE_PATH):packages/aura-core/src uv run pytest core/tests/ -v
 	# Run telegram-bot tests with isolated path to avoid 'src' collision
-	PYTHONPATH=$(TG_PATH) uv run pytest synapses/telegram-bot/tests/ -v
+	PYTHONPATH=$(TG_PATH):packages/aura-core/src uv run pytest synapses/telegram-bot/tests/ -v
 
 # Run tests with coverage report
 test-cov:
-	PYTHONPATH=$(CORE_PATH) uv run pytest core/tests/ -v --cov=core/src --cov-report=term-missing
+	PYTHONPATH=$(CORE_PATH):packages/aura-core/src uv run pytest core/tests/ -v --cov=core/src --cov-report=term-missing
 
 # Run tests with verbose output
 test-verbose:
-	PYTHONPATH=$(CORE_PATH) uv run pytest core/tests/ -vv -s
+	PYTHONPATH=$(CORE_PATH):packages/aura-core/src uv run pytest core/tests/ -vv -s
 
 # --- 2. BUILD ---
 build: generate build-tg
