+++ agents/bee-keeper/src/hive/connector/proteins/vcs/__init__.py
@@ -1,3 +1,3 @@
-from .skill import VCS_Skill
+from .main import VCS_Skill
 
 __all__ = ["VCS_Skill"]
