+++ agents/bee-keeper/src/hive/connector/proteins/vcs/manifest.yaml
@@ -2,4 +2,3 @@ name: vcs
 description: Version Control System Protein for GitHub interactions.
 capabilities:
   - post_comment
-  - reply_to_comment
