+++ agents/bee-keeper/src/hive/connector/proteins/vcs/schema.py
@@ -7,10 +7,3 @@ class CommentParams(BaseModel):
     issue_number: Optional[int] = None
     commit_sha: Optional[str] = None
     body: str = ""
-
-
-class ReplyParams(BaseModel):
-    repo: str
-    pull_number: int
-    comment_id: int
-    body: str
