+++ core/src/hive/proteins/guard/logic.py
