+++ core/src/hive/proteins/guard/skill.py
@@ -5,7 +5,7 @@ from aura_core import Observation, SkillProtocol
 
 from config.policy import SafetySettings
 
-from .engine import OutputGuard, SafetyViolation
+from .logic import OutputGuard, SafetyViolation
 from .schema import SafePriceParams, ValidationParams
 
 logger = logging.getLogger(__name__)
@@ -16,23 +16,18 @@ class GuardSkill(
 ):
     """
     Guard Protein: Handles safety validation and safe price calculation.
+    Crystallized: Implementation logic moved to logic.py.
     """
 
     def __init__(self) -> None:
         self.settings: SafetySettings | None = None
         self.provider: OutputGuard | None = None
-        self._capabilities = {
-            "validate_decision": self._validate_decision,
-            "validate_margin": self._validate_decision,
-            "validate_floor": self._validate_decision,
-            "get_safe_price": self._get_safe_price,
-        }
 
     def get_name(self) -> str:
         return "guard"
 
     def get_capabilities(self) -> list[str]:
-        return list(self._capabilities.keys())
+        return ["validate_decision", "get_safe_price"]
 
     def bind(self, settings: SafetySettings, provider: OutputGuard) -> None:
         self.settings = settings
@@ -44,13 +39,18 @@ class GuardSkill(
     async def execute(self, intent: str, params: dict[str, Any]) -> Observation:
         if not self.provider:
             return Observation(success=False, error="provider_not_initialized")
-
-        handler = self._capabilities.get(intent)
-        if not handler:
-            return Observation(success=False, error=f"Unknown intent: {intent}")
-
         try:
-            return await handler(params)
+            if intent in ["validate_decision", "validate_margin", "validate_floor"]:
+                p = ValidationParams(**params)
+                self.provider.validate_decision(p.decision, p.context)
+                return Observation(success=True)
+
+            elif intent == "get_safe_price":
+                p_safe = SafePriceParams(**params)
+                price = self.provider.calculate_safe_price(p_safe.context, p_safe.reason)
+                return Observation(success=True, data={"safe_price": price})
+
+            return Observation(success=False, error=f"Unknown intent: {intent}")
         except SafetyViolation as e:
             err_msg = str(e)
             code = "SAFETY_VIOLATION"
@@ -59,7 +59,6 @@ class GuardSkill(
             elif "floor" in err_msg.lower():
                 code = "FLOOR_PRICE_VIOLATION"
 
-            assert self.provider is not None
             safe_p = self.provider.calculate_safe_price(params.get("context", {}), code)
             return Observation(
                 success=False,
@@ -67,17 +66,4 @@ class GuardSkill(
                 data={"error_code": code, "safe_price": safe_p},
             )
         except Exception as e:
-            logger.error(f"Guard skill error: {e}")
             return Observation(success=False, error=str(e))
-
-    async def _validate_decision(self, params: dict[str, Any]) -> Observation:
-        assert self.provider is not None
-        p = ValidationParams(**params)
-        self.provider.validate_decision(p.decision, p.context)
-        return Observation(success=True)
-
-    async def _get_safe_price(self, params: dict[str, Any]) -> Observation:
-        assert self.provider is not None
-        p_safe = SafePriceParams(**params)
-        price = self.provider.calculate_safe_price(p_safe.context, p_safe.reason)
-        return Observation(success=True, data={"safe_price": price})
