+++ core/src/hive/proteins/pulse/schema.py
@@ -5,14 +5,12 @@ from pydantic import BaseModel
 
 class EventParams(BaseModel):
     """Deprecated: Use typed event params instead."""
-
     topic: str
     payload: dict[str, Any]
 
 
 class NegotiationEventParams(BaseModel):
     """Params for emit_negotiation intent."""
-
     session_token: str
     action: str  # accept, counter, reject
     price: float
@@ -22,7 +20,6 @@ class NegotiationEventParams(BaseModel):
 
 class VitalsEventParams(BaseModel):
     """Params for emit_vitals intent."""
-
     service: str
     cpu_usage: float
     memory_usage: float
@@ -31,7 +28,6 @@ class VitalsEventParams(BaseModel):
 
 class AlertEventParams(BaseModel):
     """Params for emit_alert intent."""
-
     severity: str  # info, warning, error, critical
     message: str
     source: str
@@ -39,7 +35,6 @@ class AlertEventParams(BaseModel):
 
 class AuditEventParams(BaseModel):
     """Params for emit_audit intent."""
-
     repo_name: str
     is_pure: bool
     heresies: list[str] = []
