+++ core/src/hive/proteins/reasoning/skill.py
@@ -18,6 +18,7 @@ class ReasoningSkill(
 ):
     """
     Reasoning Protein: Handles LLM logic, DSPy negotiation, and embeddings.
+    Standardized following the Crystalline Protein Standard and Enzyme pattern.
     """
 
     def __init__(self) -> None:
@@ -25,16 +26,12 @@ class ReasoningSkill(
         self.provider: dict[str, Any] | None = None
         self.negotiator: Any = None
         self._embed_model: Any = None
-        self._capabilities = {
-            "negotiate": self._negotiate,
-            "generate_embedding": self._generate_embedding,
-        }
 
     def get_name(self) -> str:
         return "reasoning"
 
     def get_capabilities(self) -> list[str]:
-        return list(self._capabilities.keys())
+        return ["negotiate", "generate_embedding"]
 
     def bind(self, settings: LLMSettings, provider: dict[str, Any]) -> None:
         self.settings = settings
@@ -60,48 +57,48 @@ class ReasoningSkill(
         return True
 
     async def execute(self, intent: str, params: dict[str, Any]) -> Observation:
-        handler = self._capabilities.get(intent)
-        if not handler:
-            return Observation(success=False, error=f"Unknown intent: {intent}")
-
         try:
-            return await handler(params)
+            if intent == "negotiate":
+                if not self.negotiator:
+                    return Observation(success=False, error="negotiator_not_ready")
+                p_neg = NegotiationParams(**params)
+
+                def call() -> dict[str, Any]:
+                    from typing import cast
+
+                    neg = cast(Any, self.negotiator)
+                    return cast(
+                        dict[str, Any],
+                        neg(
+                            input_bid=p_neg.bid,
+                            context=p_neg.context,
+                            history=p_neg.history,
+                        ),
+                    )
+
+                result = await asyncio.to_thread(call)
+                data = {
+                    "action": result["action"]["action"],
+                    "price": result["action"]["price"],
+                    "message": result["action"]["message"],
+                    "thought": result.get("thought", ""),
+                    "metadata": result.get("metadata", {}),
+                }
+                return Observation(
+                    success=True, data=NegotiationResult(**data).model_dump()
+                )
+
+            elif intent == "generate_embedding":
+                if not self._embed_model:
+                    return Observation(success=False, error="embed_model_not_ready")
+                p_emb = EmbeddingParams(**params)
+
+                emb = await asyncio.to_thread(
+                    generate_embedding, p_emb.text, self._embed_model
+                )
+                return Observation(success=True, data=emb)
+
+            return Observation(success=False, error=f"Unknown intent: {intent}")
         except Exception as e:
             logger.error(f"Reasoning skill error: {e}")
             return Observation(success=False, error=str(e))
-
-    async def _negotiate(self, params: dict[str, Any]) -> Observation:
-        if not self.negotiator:
-            return Observation(success=False, error="negotiator_not_ready")
-        p_neg = NegotiationParams(**params)
-
-        def call() -> dict[str, Any]:
-            from typing import cast
-
-            neg = cast(Any, self.negotiator)
-            return cast(
-                dict[str, Any],
-                neg(
-                    input_bid=p_neg.bid,
-                    context=p_neg.context,
-                    history=p_neg.history,
-                ),
-            )
-
-        result = await asyncio.to_thread(call)
-        data = {
-            "action": result["action"]["action"],
-            "price": result["action"]["price"],
-            "message": result["action"]["message"],
-            "thought": result.get("thought", ""),
-            "metadata": result.get("metadata", {}),
-        }
-        return Observation(success=True, data=NegotiationResult(**data).model_dump())
-
-    async def _generate_embedding(self, params: dict[str, Any]) -> Observation:
-        if not self._embed_model:
-            return Observation(success=False, error="embed_model_not_ready")
-        p_emb = EmbeddingParams(**params)
-
-        emb = await asyncio.to_thread(generate_embedding, p_emb.text, self._embed_model)
-        return Observation(success=True, data=emb)
