+++ core/src/main.py
@@ -13,17 +13,36 @@ from hive.metabolism.logging_config import (
     configure_logging,
     get_logger,
 )
+from hive.proteins.telemetry.prometheus import init_telemetry
 from hive.proto.aura.negotiation.v1 import negotiation_pb2, negotiation_pb2_grpc
 from opentelemetry import trace
+from opentelemetry.instrumentation.grpc import GrpcInstrumentorServer
+from opentelemetry.instrumentation.langchain import LangchainInstrumentor
+from prometheus_client import start_http_server
 
 from config import settings
 
 if TYPE_CHECKING:
     from hive.metabolism import MetabolicLoop
 
-# 1. Configure structured logging
+# Configure structured logging on startup
 configure_logging(log_level=settings.server.log_level)
 logger = get_logger("core")
+
+# Initialize OpenTelemetry tracing via Telemetry Protein's prometheus implementation
+service_name = settings.server.otel_service_name
+tracer = init_telemetry(service_name, str(settings.server.otel_exporter_otlp_endpoint))
+logger.info(
+    "telemetry_initialized",
+    service_name=service_name,
+    endpoint=str(settings.server.otel_exporter_otlp_endpoint),
+)
+
+# Instrument gRPC server for distributed tracing
+GrpcInstrumentorServer().instrument()
+
+# Instrument LangChain for LLM call tracing
+LangchainInstrumentor().instrument()
 
 # gRPC metadata key for request_id
 REQUEST_ID_METADATA_KEY = "x-request-id"
@@ -43,7 +62,7 @@ class NegotiationService(negotiation_pb2_grpc.NegotiationServiceServicer):
 
     def __init__(
         self,
-        metabolism: "MetabolicLoop" | None = None,
+        metabolism: 'MetabolicLoop' | None = None,
         market_service: Any = None,
     ) -> None:
         self.metabolism = metabolism
@@ -52,8 +71,12 @@ class NegotiationService(negotiation_pb2_grpc.NegotiationServiceServicer):
     async def Negotiate(
         self, request: Any, context: Any
     ) -> negotiation_pb2.NegotiateResponse:
-        """Main metabolic loop for negotiation."""
+        """
+        Main metabolic loop for negotiation:
+        Signal -> A -> T -> Membrane -> C -> G
+        """
         if not self.metabolism:
+            logger.warning("negotiate_called_before_initialization")
             context.set_code(grpc.StatusCode.UNAVAILABLE)
             context.set_details("Metabolism is still initializing")
             return negotiation_pb2.NegotiateResponse()
@@ -115,6 +138,11 @@ class NegotiationService(negotiation_pb2_grpc.NegotiationServiceServicer):
                 "generate_embedding", {"text": request.query}
             )
             if not embed_obs.success:
+                logger.error(
+                    "embedding_generation_failed",
+                    query=request.query,
+                    error=embed_obs.error,
+                )
                 context.set_code(grpc.StatusCode.INTERNAL)
                 context.set_details(f"Failed to generate embedding: {embed_obs.error}")
                 return negotiation_pb2.SearchResponse()
@@ -129,6 +157,7 @@ class NegotiationService(negotiation_pb2_grpc.NegotiationServiceServicer):
             )
 
             if not obs.success:
+                logger.error("vector_search_failed", error=obs.error)
                 context.set_code(grpc.StatusCode.INTERNAL)
                 context.set_details(f"Search failed: {obs.error}")
                 return negotiation_pb2.SearchResponse()
@@ -144,10 +173,11 @@ class NegotiationService(negotiation_pb2_grpc.NegotiationServiceServicer):
                 for item in obs.data
             ]
 
+            logger.info("search_completed", result_count=len(response_items))
             return negotiation_pb2.SearchResponse(results=response_items)
 
         except Exception as e:
-            logger.error("search_error", error=str(e), exc_info=True)
+            logger.error("search_error", error=str(e))
             context.set_code(grpc.StatusCode.INTERNAL)
             context.set_details(str(e))
             return negotiation_pb2.SearchResponse()
@@ -184,7 +214,9 @@ class NegotiationService(negotiation_pb2_grpc.NegotiationServiceServicer):
             bind_request_id(request_id)
 
         try:
+            # Feature toggle check
             if not settings.crypto.enabled or not self.market_service:
+                logger.warning("crypto_disabled", deal_id=request.deal_id)
                 context.set_code(grpc.StatusCode.UNIMPLEMENTED)
                 context.set_details("Crypto payments not enabled")
                 return negotiation_pb2.CheckDealStatusResponse(status="NOT_FOUND")
@@ -193,13 +225,23 @@ class NegotiationService(negotiation_pb2_grpc.NegotiationServiceServicer):
             try:
                 uuid.UUID(request.deal_id)
             except ValueError:
+                logger.warning("invalid_deal_id", deal_id=request.deal_id)
                 context.set_code(grpc.StatusCode.INVALID_ARGUMENT)
                 context.set_details("Invalid deal_id format")
                 return negotiation_pb2.CheckDealStatusResponse(status="NOT_FOUND")
 
+            logger.info("check_deal_status_started", deal_id=request.deal_id)
+
+            # Check payment status via MarketService
             response = cast(
                 negotiation_pb2.CheckDealStatusResponse,
                 await self.market_service.check_status(deal_id=request.deal_id),
+            )
+
+            logger.info(
+                "check_deal_status_completed",
+                deal_id=request.deal_id,
+                status=response.status,
             )
             return response
 
@@ -221,7 +263,7 @@ class NegotiationService(negotiation_pb2_grpc.NegotiationServiceServicer):
 async def serve() -> None:
     from grpc_health.v1 import health
 
-    # 1. Initialize gRPC Server
+    # 1. Initialize gRPC Server early
     server = grpc.aio.server(
         futures.ThreadPoolExecutor(max_workers=settings.server.grpc_max_workers)
     )
@@ -230,29 +272,35 @@ async def serve() -> None:
     health_servicer = health.HealthServicer()
     health_pb2_grpc.add_HealthServicer_to_server(health_servicer, server)
 
-    # 3. Register Negotiation Service
+    # 3. Register Negotiation Service with placeholders
     negotiation_service = NegotiationService()
     negotiation_pb2_grpc.add_NegotiationServiceServicer_to_server(
         negotiation_service, server
     )
 
-    # 4. Start the server early
+    # 4. Bind and Start the server
     server.add_insecure_port(f"[::]:{settings.server.port}")
     await server.start()
     logger.info("server_started_early", port=settings.server.port)
 
-    # 5. Initialize the "Cell" (The HiveCell)
+    # 5. Start Prometheus metrics server
+    try:
+        start_http_server(9091)
+    except Exception as e:
+        logger.error("metrics_server_failed", error=str(e))
+
+    # 6. Initialize the "Cell" (The Cortex)
     cell = HiveCell(settings)
     metabolism = await cell.build_organism()
 
-    # 6. Wire fully initialized components
+    # 7. Wire fully initialized components
     negotiation_service.metabolism = metabolism
     negotiation_service.market_service = cell.market_service
 
     # Set health to SERVING once DB/Metabolism is up
     health_servicer.set("", health_pb2.HealthCheckResponse.SERVING)
 
-    # 7. Start Heartbeat Deal (Honey Stimulus)
+    # 8. Start Heartbeat Deal (Honey Stimulus)
     async def heartbeat_deal_loop() -> None:
         await asyncio.sleep(60)
         while True:
@@ -265,8 +313,7 @@ async def serve() -> None:
                     item = obs.data
                     mock_signal = negotiation_pb2.NegotiateRequest(
                         item_id=item["id"],
-                        bid_amount=item["base_price"]
-                        * settings.heartbeat.bid_multiplier,
+                        bid_amount=item["base_price"] * settings.heartbeat.bid_multiplier,
                         currency_code="USD",
                         agent=negotiation_pb2.AgentIdentity(
                             did=settings.heartbeat.agent_did,
