+++ core/tests/test_membrane.py
@@ -9,7 +9,7 @@ async def test_membrane_rule1_floor_price_override():
     """
     Rule 1: If price < floor_price, override to counter-offer at floor_price + 5%.
     """
-    from hive.proteins.guard.engine import OutputGuard
+    from hive.proteins.guard.logic import OutputGuard
 
     from config.policy import SafetySettings
 
@@ -43,7 +43,7 @@ async def test_membrane_rule2_data_leak_prevention():
     """
     Rule 2: Block any response containing "floor_price" in the human message.
     """
-    from hive.proteins.guard.engine import OutputGuard
+    from hive.proteins.guard.logic import OutputGuard
 
     from config.policy import SafetySettings
 
@@ -78,7 +78,7 @@ async def test_membrane_combined_violations():
     """
     Test both Rule 1 and Rule 2 triggered at once.
     """
-    from hive.proteins.guard.engine import OutputGuard
+    from hive.proteins.guard.logic import OutputGuard
 
     from config.policy import SafetySettings
